+++ CHANGELOG.md
@@ -1,7 +1,6 @@
-***Version 1.0.0.2018* ** *(TBD)*
+***Version 1.0.0.2018* ** *(December 16, 2018)*
 - Fixed a localization bug in the info screen.
-- Disable "Audible Ticks" while fetching music.
-- Fixed a bug, where the vibrate switch wasn't being properly disabled while fetching music.
+- More accessibility work.
 
 ***Version 1.0.0.2017* ** *(December 15, 2018)*
 - Fixed a crash (my bad) when you open the info screen.